+++ agentic_blog_bot/__init__.py
@@ -6,17 +6,17 @@ BASE_DIR = Path(__file__).resolve().parent.parent
 if str(BASE_DIR) not in sys.path:
     sys.path.insert(0, str(BASE_DIR))
 
-_modules = [
-    "agent_tools",
-    "github_utils",
-    "config",
-    "notifier",
-    "linkedin_poster",
-]
+_modules_map = {
+    "agent_tools": "authorship.agent_tools",
+    "github_utils": "publishing.github_utils",
+    "config": "config",
+    "notifier": "publishing.notifier",
+    "linkedin_poster": "publishing.linkedin_poster",
+}
 
-for _m in _modules:
-    module = importlib.import_module(_m)
-    sys.modules[f"agentic_blog_bot.{_m}"] = module
-    setattr(sys.modules[__name__], _m, module)
+for name, path in _modules_map.items():
+    module = importlib.import_module(path)
+    sys.modules[f"agentic_blog_bot.{name}"] = module
+    setattr(sys.modules[__name__], name, module)
 
-__all__ = _modules
+__all__ = list(_modules_map.keys())
