+++ agentic_blog_bot/agent_tools.py
@@ -1,7 +1,6 @@
 """Proxy module for backwards compatibility."""
-
 from importlib import import_module
 import sys
-
-_module = import_module("agent_tools")
+_module = import_module("authorship.agent_tools")
 sys.modules[__name__] = _module
+
