+++ agentic_blog_bot/github_utils.py
@@ -1,7 +1,6 @@
 """Proxy module for backwards compatibility."""
-
 from importlib import import_module
 import sys
-
-_module = import_module("github_utils")
+_module = import_module("publishing.github_utils")
 sys.modules[__name__] = _module
+
