+++ agentic_blog_bot/linkedin_poster.py
@@ -1,7 +1,6 @@
 """Proxy module for backwards compatibility."""
-
 from importlib import import_module
 import sys
-
-_module = import_module("linkedin_poster")
+_module = import_module("publishing.linkedin_poster")
 sys.modules[__name__] = _module
+
