+++ authorship/__init__.py
@@ -1,0 +1,2 @@
+"""LLM-driven writing helpers and Discord bot."""
+
