+++ curation/__init__.py
@@ -1,0 +1,2 @@
+"""Tools for scraping and indexing chat conversations."""
+
