+++ curation/check_auth.py
