+++ publishing/__init__.py
@@ -1,0 +1,2 @@
+"""Publishing utilities for GitHub and LinkedIn."""
+
