+++ test_github_utils.py
@@ -1,4 +1,8 @@
-from github_utils import compute_diff_sha, should_skip_review, update_cached_sha
+from publishing.github_utils import (
+    compute_diff_sha,
+    should_skip_review,
+    update_cached_sha,
+)
 
 def test_compute_diff_sha_consistency():
     diff = "file.md:\n- old\n+ new"
@@ -13,7 +17,7 @@ def test_should_skip_review(tmp_path, monkeypatch):
     cache_file = tmp_path / "pr_shas.json"
     cache_file.write_text(f'{{"{pr_number}": "{test_sha}"}}')
 
-    monkeypatch.setattr("github_utils.PR_SHAS_FILE", cache_file)
+    monkeypatch.setattr("publishing.github_utils.PR_SHAS_FILE", cache_file)
 
     assert should_skip_review(pr_number, "abc123")
     assert not should_skip_review(pr_number, "xyz789")
@@ -22,7 +26,7 @@ def test_update_cached_sha_roundtrip(tmp_path, monkeypatch):
     pr_number = 100
     sha = "def456"
     cache_file = tmp_path / "pr_shas.json"
-    monkeypatch.setattr("agentic_blog_bot.github_utils.PR_SHAS_FILE", cache_file)
+    monkeypatch.setattr("publishing.github_utils.PR_SHAS_FILE", cache_file)
 
     update_cached_sha(pr_number, sha)
     try:
